refactor: share the thinking-aware stream between providers
A provider that reasons has two ways to say so and may use either. The loop
zhipu.ts grew for issue #116 handles both — `reasoning_content` as a separate
field, and `<think>` tags buffered across chunk boundaries — and nothing in it
is specific to Zhipu. Lift it into utils as streamWithThinking so the next
provider that reasons does not have to rediscover the same two cases.

No behaviour change intended, and none observed: glm-z1-flash and glm-4.6 still
open a callout, glm-4-plus still answers without gaining a stray one.
